+++ slack-market-game/api/endpoints.rb
@@ -1,0 +1,3 @@
+require 'slack-market-game/api/endpoints/teams_endpoint'
+require 'slack-market-game/api/endpoints/status_endpoint'
+require 'slack-market-game/api/endpoints/root_endpoint'