+++ slack-market-game/api/presenters.rb
@@ -1,0 +1,9 @@
+require 'roar/representer'
+require 'roar/json'
+require 'roar/json/hal'
+
+require 'slack-market-game/api/presenters/paginated_presenter'
+require 'slack-market-game/api/presenters/status_presenter'
+require 'slack-market-game/api/presenters/team_presenter'
+require 'slack-market-game/api/presenters/teams_presenter'
+require 'slack-market-game/api/presenters/root_presenter'