+++ slack-market-game/models.rb
@@ -1,0 +1,1 @@
+require 'slack-market-game/models/team'