+++ slack-market-game/service.rb
@@ -1,4 +1,4 @@
-module SlackBotServer
+module SlackMarketGame
   class Service
     LOCK = Mutex.new
     @services = {}
@@ -8,7 +8,7 @@ module SlackBotServer
         fail 'Token already known.' if @services.key?(team.token)
         EM.next_tick do
           logger.info "Starting team #{team}."
-          server = SlackBotServer::Server.new(team: team)
+          server = SlackMarketGame::Server.new(team: team)
           LOCK.synchronize do
             @services[team.token] = server
           end